+++ src/main/scala/com/sparkfits/FitsFileInputFormat.scala
@@ -16,10 +16,13 @@
 package com.sparkfits
 
 import org.apache.hadoop.io.LongWritable
-import org.apache.hadoop.mapreduce.{InputSplit, JobContext, RecordReader, TaskAttemptContext}
+import org.apache.hadoop.mapreduce.InputSplit
+import org.apache.hadoop.mapreduce.JobContext
+import org.apache.hadoop.mapreduce.RecordReader
+import org.apache.hadoop.mapreduce.TaskAttemptContext
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat
+
 import org.apache.spark.sql.Row
-import com.sparkfits.FitsRecordReader
 
 private[sparkfits] object FitsFileInputFormat {
   /** Property name to set in Hadoop JobConfs for record length */
