+++ src/main/scala/com/sparkfits/old/trash.scala
